Add EOL Cohort ORM Example
We added a specialized ORM version of our CohortTable for the EOL
task, which uses slightly different logic than what we had already
implemented.

This included creating the EOL cohort class, creating a notebook
to test it for the deep model training example, and updating the
config file.

We'll use this new class to perform benchmarks against the original
EOL cohort definition, to check for bugs but also to compare speed
and memory performance.